+++ app/Http/Controllers/UserController.php
@@ -20,30 +20,40 @@ class UserController extends BaseController
     }
     public function index()
     {
-        $this->setPageData('User','User','fas fa-users');
-        $roles = $this->role->index();
-        return view('user.index',compact('roles'));
+        if(permission('user-access')){
+            $this->setPageData('User','User','fas fa-users');
+            $roles = $this->role->index();
+            return view('user.index',compact('roles'));
+        }else{
+            return $this->unauthorized_access_blocked();
+        }
     }
 
     public function getDatatableData(Request $request)
     {
-        if($request->ajax()){
-            $output = $this->service->getDatatableData($request);
-        }else{
-            $output = ['status' => 'error', 'message' => 'Unauthorized action blocked!'];
-        }
+        if(permission('user-access')){
+            if($request->ajax()){
+                $output = $this->service->getDatatableData($request);
+            }else{
+                $output = ['status' => 'error', 'message' => 'Unauthorized action blocked!'];
+            }
 
-        return response()->json($output);
+            return response()->json($output);
+        }
     }
 
     public function storeOrUpdateData(UserFormRequest $request)
     {
         if($request->ajax()){
-            $Result = $this->service->storeOrUpdateData($request);
-            if($Result){
-                return $this->response_json('success','Data has been saved successfully',null,200);
+            if(permission('user-access') || permission('user-edit')){
+                $Result = $this->service->storeOrUpdateData($request);
+                if($Result){
+                    return $this->response_json('success','Data has been saved successfully',null,200);
+                }else{
+                    return $this->response_json('error','Data cannot be saved',null,204);
+                }
             }else{
-                return $this->response_json('error','Data cannot be saved',null,204);
+                return $this->response_json($status='error',$message='Unauthorized Access Blocked',$data=null,$response_code=401);
             }
         }else{
             return $this->response_json('error',null,null,401);
@@ -53,11 +63,15 @@ class UserController extends BaseController
     public function edit(Request $request)
     {
         if($request->ajax()){
-            $data = $this->service->edit($request);
-            if($data->count()){
-                return $this->response_json('success',null,$data,201);
+            if (permission('user-edit')){
+                $data = $this->service->edit($request);
+                if($data->count()){
+                    return $this->response_json('success',null,$data,201);
+                }else{
+                    return $this->response_json('error','No data found',null,204);
+                }
             }else{
-                return $this->response_json('error','No data found',null,204);
+                return $this->response_json($status='error',$message='Unauthorized Access Blocked',$data=null,$response_code=401);
             }
         }else{
             return $this->response_json('error',null,null,401);
@@ -67,19 +81,25 @@ class UserController extends BaseController
     public function show(Request $request)
     {
         if($request->ajax()){
-            $user = $this->service->edit($request);
-            return view('user.details',compact('user'))->render();
+            if (permission('user-view')){
+                $user = $this->service->edit($request);
+                return view('user.details',compact('user'))->render();
+            }
         }
     }
 
     public function delete(Request $request)
     {
         if($request->ajax()){
-            $Result = $this->service->delete($request);
-            if($Result){
-                return $this->response_json('success','Data has been deleted successfully',null,200);
+            if(permission('user-delete')){
+                $Result = $this->service->delete($request);
+                if($Result){
+                    return $this->response_json('success','Data has been deleted successfully',null,200);
+                }else{
+                    return $this->response_json('error','Data cannot be deleted',null,204);
+                }
             }else{
-                return $this->response_json('error','Data cannot be deleted',null,204);
+                return $this->response_json($status='error',$message='Unauthorized Access Blocked',$data=null,$response_code=401);
             }
         }else{
             return $this->response_json('error',null,null,401);
@@ -89,11 +109,15 @@ class UserController extends BaseController
     public function bulkDelete(Request $request)
     {
         if($request->ajax()){
-            $Result = $this->service->bulkDelete($request);
-            if($Result){
-                return $this->response_json('success','Data has been deleted successfully',null,200);
+            if(permission('user-bulk-delete')){
+                $Result = $this->service->bulkDelete($request);
+                if($Result){
+                    return $this->response_json('success','Data has been deleted successfully',null,200);
+                }else{
+                    return $this->response_json('error','Data cannot be deleted',null,204);
+                }
             }else{
-                return $this->response_json('error','Data cannot be deleted',null,204);
+                return $this->response_json($status='error',$message='Unauthorized Access Blocked',$data=null,$response_code=401);
             }
         }else{
             return $this->response_json('error',null,null,401);
@@ -103,11 +127,15 @@ class UserController extends BaseController
     public function changeStatus(Request $request)
     {
         if($request->ajax()){
-            $Result = $this->service->changeStatus($request);
-            if($Result){
-                return $this->response_json('success','Status has been changed successfully',null,200);
+            if (permission('user-edit')) {
+                $Result = $this->service->changeStatus($request);
+                if($Result){
+                    return $this->response_json('success','Status has been changed successfully',null,200);
+                }else{
+                    return $this->response_json('error','Status cannot be changed!',null,204);
+                }
             }else{
-                return $this->response_json('error','Status cannot be changed!',null,204);
+                return $this->response_json($status='error',$message='Unauthorized Access Blocked',$data=null,$response_code=401);
             }
         }else{
             return $this->response_json('error',null,null,401);
